Moved some stuff around and fixed some bugs. Naming still needs to tweaking.
