+++ UI/Web/src/app/chapter-detail/chapter-detail.component.html
@@ -85,7 +85,7 @@
               </div>
 
               <div class="col-auto ms-2 d-none d-md-block">
-                <app-download-button [entity]="chapterValue" entityType="chapter" />
+                <app-download-button [entity]="chapterValue" entityType="chapter" [libraryId]="libraryId()" [seriesId]="seriesId()" />
               </div>
 
             </div>
