+++ UI/Web/src/app/nav/_components/download-queue-widget/download-queue-widget.component.html
@@ -1,7 +1,7 @@
 <ng-container *transloco="let t; prefix:'download-queue-widget'">
   @if (downloadService.hasActiveDownloads() || downloadService.completedTodayCount() > 0 || downloadService.olderCompletedCount() > 0 || downloadService.failedItems().length > 0) {
     <div class="nav-item me-2">
-      <button class="btn btn-icon position-relative" (click)="openDownloadQueue()" [title]="t('title')">
+      <button class="btn btn-icon position-relative" (click)="openDownloadQueue()" [title]="t('title')" [attr.aria-label]="t('title')">
         <i class="fa fa-download nav"
            [class.colored]="downloadService.hasActiveDownloads() && !downloadService.isPaused()"
            [class.text-warning]="downloadService.isPaused() && downloadService.queuedItems().length > 0"
