+++ UI/Web/src/app/volume-detail/volume-detail.component.html
@@ -84,7 +84,7 @@
             </div>
 
             <div class="col-auto ms-2 d-none d-md-block">
-              <app-download-button [entity]="volume()" entityType="volume" />
+              <app-download-button [entity]="volume()" entityType="volume" [libraryId]="libraryId()" [seriesId]="seriesId()" />
             </div>
 
           </div>
